color-picker/challenge/requirements.txt
flask
requests
flask-csp
google-auth
oauth2client